requirements.txt
autopep8==1.5.7
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
decorator==5.0.9
dialogflow==1.1.0
feedparser==6.0.8
Flask==2.0.1
fuzzywuzzy==0.18.0
gnewsclient==1.12
google-api-core==1.30.0
google-auth==1.32.1
google-cloud-dialogflow==2.5.0
googleapis-common-protos==1.53.0
grpcio==1.38.1
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyparsing==2.4.7
python-telegram-bot==12.8
pytz==2021.1
requests==2.25.1
rsa==4.7.2
sgmllib3k==1.0.0
six==1.16.0
toml==0.10.2
tornado==6.1
urllib3==1.26.6
Werkzeug==2.0.1
